Prevent accidental changes to a dictionary if the dictionary displayed on the screen is not the "current_dictionary".
This would only happen if you had Lexiconga open in two windows/tabs and changed your dictionary in one window/tab without refreshing the other, then making edits to the un-refreshed window/tab. Now it also sends which dictionary to change and makes sure that dictionary is in the user's session dictionaries list so it updates the correct dictionary's words.